+++ app/src/main/java/com/concordium/wallet/ui/cis2/SendTokenActivity.kt
@@ -22,7 +22,7 @@ import com.concordium.wallet.databinding.ActivitySendTokenBinding
 import com.concordium.wallet.ui.base.BaseActivity
 import com.concordium.wallet.ui.cis2.SendTokenViewModel.Companion.SEND_TOKEN_DATA
 import com.concordium.wallet.ui.recipient.recipientlist.RecipientListActivity
-import com.concordium.wallet.ui.recipient.scanqr.ScanQRActivity
+import com.concordium.wallet.ui.scanqr.ScanQRActivity
 import com.concordium.wallet.ui.transaction.sendfunds.AddMemoActivity
 import com.concordium.wallet.util.CBORUtil
 import com.concordium.wallet.util.KeyboardUtil
@@ -200,12 +200,14 @@ class SendTokenActivity : BaseActivity(R.layout.activity_send_token, R.string.ci
         }
 
     private val getResultScanQr =
-        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) {
-            if (it.resultCode == Activity.RESULT_OK) {
-                it.data?.getStringExtra(ScanQRActivity.EXTRA_BARCODE)?.let { barcode ->
-                    binding.receiver.setText(barcode)
+        registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
+            result
+                .takeIf { it.resultCode == Activity.RESULT_OK }
+                ?.data?.extras
+                ?.let(ScanQRActivity::getScannedQrContent)
+                ?.also { scannedQrContent ->
+                    binding.receiver.setText(scannedQrContent)
                 }
-            }
         }
 
     private fun clearMemo() =