+++ app/src/main/java/com/concordium/wallet/ui/welcome/WelcomePromoActivity.kt
@@ -8,6 +8,7 @@ import androidx.fragment.app.commit
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.get
 import com.concordium.wallet.App
+import com.concordium.wallet.BuildConfig
 import com.concordium.wallet.R
 import com.concordium.wallet.databinding.ActivityWelcomePromoBinding
 import com.concordium.wallet.ui.auth.passcode.PasscodeSetupActivity
@@ -57,10 +58,17 @@ class WelcomePromoActivity :
             App.appCore.tracker.welcomeHomeScreen()
         }
 
+        binding.bottomNavigationView.menu.findItem(R.id.menuitem_news).isVisible =
+            BuildConfig.SHOW_NEWSFEED
         binding.bottomNavigationView.setOnItemSelectedListener { item ->
             when (item.itemId) {
                 R.id.menuitem_accounts -> {
                     showAccounts()
+                    true
+                }
+
+                R.id.menuitem_news -> {
+                    showNews()
                     true
                 }
 
@@ -123,6 +131,11 @@ class WelcomePromoActivity :
             disallowAddToBackStack()
             replace(R.id.fragment_container, WelcomePromoAccountsFragment())
         }
+
+    private fun showNews() = supportFragmentManager.commit {
+        disallowAddToBackStack()
+        replace(R.id.fragment_container, WelcomePromoNewsOverviewFragment())
+    }
 
     private fun showMore() = supportFragmentManager.commit {
         disallowAddToBackStack()